Calculating the area of a triangle
In this python code, we can calculate the area of a triangle using Heron's formula, and you can also determine the height corresponding to the longest side of the triangle. I'll also add some graphical options, such as the Turtle library. Just input three numbers, and then you will get the area and height.